refractor(#77): Fix comments
