+++ frontend/public/locales/en/common.json
@@ -1,5 +1,6 @@
 {
   "home": "Home",
+  "dsek": "The Guild",
   "documents": "Documents",
   "news": "News",
   "events": "Events",
