+++ frontend/public/locales/en/mandate.json
@@ -3,5 +3,5 @@
   "positions": "Positions",
   "mandateError": "No mandates were found for this year.",
   "current mandates": "Current Mandates",
-  "vakant": "V.A.Kant"
+  "vakant": "V.A.Cant"
 }
