+++ frontend/public/locales/sv/common.json
@@ -1,5 +1,6 @@
 {
   "home": "Hem",
+  "dsek": "Sektionen",
   "documents": "Dokument",
   "news": "Nyheter",
   "events": "Events",
@@ -7,7 +8,7 @@
   "create_new_event": "Skapa nytt event",
   "calendar": "Kalender",
   "booking": "Bokning",
-  "mandates": "Mandat",
+  "mandates": "Funktionärer",
   "archive": "Arkiv",
   "menu": "Meny",
   "regulations": "Reglemente",
