+++ frontend/public/locales/sv/mandate.json
@@ -3,5 +3,5 @@
   "positions": "Poster",
   "mandateError": "Inga mandat hittade för det här året.",
   "current mandates": "Aktiva Funktionärer",
-  "vakant": "V.A.Cant"
+  "vakant": "V.A.Kant"
 }
